fix: systemMode primary color
