+++ packages/arex/src/hooks/index.ts
@@ -1,6 +1,7 @@
 export { default as useAuthentication } from './useAuthentication';
 export { default as useCheckChrome } from './useCheckChrome';
 export { default as useColorPrimary } from './useColorPrimary';
+export { default as useDarkMode } from './useDarkMode';
 export { default as useInit } from './useInit';
 export { default as useNavPane } from './useNavPane';
 export { default as useTrace } from './useTrace';
